CLIP_TRT/requirements.txt
torch 
torchvision 
torchaudio
wheel
tensorrt
pycuda
open_clip_torch
onnx
onnxruntime-gpu
onnxoptimizer